Progressing, all tough this is not quite correct.

+++ src/test/kotlin/org/openbroker/model/EffectiveInterestRateTest.kt
@@ -8,6 +8,6 @@ class EffectiveInterestRateTest {
     @Test
     fun testComputeEffectiveInterestRate() {
         // 9.82%
-        assertEquals(0.0982, effectiveInterestRate(10_000, 0.05, 20.0, 24, AmortizationType.ANNUITY), 0.00001)
+        assertEquals(0.0982, effectiveInterestRate(10_000, 0.05, 20, 24, AmortizationType.ANNUITY).toDouble(), 0.00001)
     }
 }
